+++ examples/widgets/list/lv_example_list_2.c
@@ -1,7 +1,10 @@
-#include <stdio.h>
+#include <stdlib.h>
+
 
 #include "../../lv_examples.h"
 #if LV_USE_LIST && LV_BUILD_EXAMPLES
+
+#define USE_MOVE_TO_INDEX
 
 static lv_obj_t* list1;
 static lv_obj_t* list2;
@@ -58,25 +61,54 @@ static void event_handler_up(lv_event_t* e)
     if ((code == LV_EVENT_CLICKED) || (code == LV_EVENT_LONG_PRESSED_REPEAT))
     {
         if (currentButton == NULL) return;
+#ifndef USE_MOVE_TO_INDEX
         lv_obj_move_up(currentButton);
+#else
+        uint32_t index = lv_obj_get_index(currentButton);
+        if (index <= 0) return;
+        lv_obj_move_to_index(currentButton, index - 1);
+#endif
+        lv_obj_scroll_to_view(currentButton, LV_ANIM_ON);
+    }
+}
+
+static void event_handler_center(lv_event_t* e)
+{
+    const lv_event_code_t code = lv_event_get_code(e);
+    if ((code == LV_EVENT_CLICKED) || (code == LV_EVENT_LONG_PRESSED_REPEAT))
+    {
+        if (currentButton == NULL) return;
+        uint32_t index = lv_obj_get_index(currentButton);
+
+        lv_obj_t* parent = lv_obj_get_parent(currentButton);
+        const uint32_t pos = lv_obj_get_child_cnt(parent) / 2;
+
+        lv_obj_move_to_index(currentButton, pos);
+
         lv_obj_scroll_to_view(currentButton, LV_ANIM_ON);
     }
 }
 
 static void event_handler_dn(lv_event_t* e)
 {
-    lv_event_code_t code = lv_event_get_code(e);
+    const lv_event_code_t code = lv_event_get_code(e);
     if ((code == LV_EVENT_CLICKED) || (code == LV_EVENT_LONG_PRESSED_REPEAT))
     {
         if (currentButton == NULL) return;
+#ifndef USE_MOVE_TO_INDEX
         lv_obj_move_down(currentButton);
+#else
+        const uint32_t index = lv_obj_get_index(currentButton);
+
+        lv_obj_move_to_index(currentButton, index + 1);
+#endif
         lv_obj_scroll_to_view(currentButton, LV_ANIM_ON);
     }
 }
 
 static void event_handler_bottom(lv_event_t* e)
 {
-    lv_event_code_t code = lv_event_get_code(e);
+    const lv_event_code_t code = lv_event_get_code(e);
     if (code == LV_EVENT_CLICKED)
     {
         if (currentButton == NULL) return;
@@ -87,28 +119,21 @@ static void event_handler_bottom(lv_event_t* e)
 
 static void event_handler_swap(lv_event_t* e)
 {
-    lv_event_code_t code = lv_event_get_code(e);
+    const lv_event_code_t code = lv_event_get_code(e);
     // lv_obj_t* obj = lv_event_get_target(e);
     if ((code == LV_EVENT_CLICKED) || (code == LV_EVENT_LONG_PRESSED_REPEAT))
     {
-        uint32_t first = 0;
-        uint32_t last = lv_obj_get_child_cnt(list1);
-        if (last > 1)
-        {
-            last--;
-            while (first < last)
+        uint32_t cnt = lv_obj_get_child_cnt(list1);
+        for (int i = 0; i < 100; i++)
+            if (cnt > 1)
             {
-                lv_obj_t* obj1 = lv_obj_get_child(list1, first);
-                lv_obj_t* obj2 = lv_obj_get_child(list1, last);
-                lv_obj_swap(obj1, obj2);
-                first++;
-                last--;
+                lv_obj_t* obj = lv_obj_get_child(list1, rand() % cnt);
+                lv_obj_move_to_index(obj, rand() % cnt);
+                if (currentButton != NULL)
+                {
+                    lv_obj_scroll_to_view(currentButton, LV_ANIM_ON);
+                }
             }
-            if (currentButton != NULL)
-            {
-                lv_obj_scroll_to_view(currentButton, LV_ANIM_ON);
-            }
-        }
     }
 }
 
@@ -122,7 +147,7 @@ void lv_example_list_2(void)
     /*Add buttons to the list*/
     lv_obj_t* btn;
     int i;
-    for (i = 0; i < 30; i++) {
+    for (i = 0; i < 15; i++) {
         btn = lv_btn_create(list1);
         lv_obj_set_width(btn, lv_pct(50));
         lv_obj_add_event_cb(btn, event_handler, LV_EVENT_CLICKED, NULL);
@@ -147,6 +172,10 @@ void lv_example_list_2(void)
 
     btn = lv_list_add_btn(list2, LV_SYMBOL_UP, "Up");
     lv_obj_add_event_cb(btn, event_handler_up, LV_EVENT_ALL, NULL);
+    lv_group_remove_obj(btn);
+
+    btn = lv_list_add_btn(list2, LV_SYMBOL_LEFT, "Center");
+    lv_obj_add_event_cb(btn, event_handler_center, LV_EVENT_ALL, NULL);
     lv_group_remove_obj(btn);
 
     btn = lv_list_add_btn(list2, LV_SYMBOL_DOWN, "Down");