+++ src/core/lv_obj_tree.c
@@ -6,6 +6,8 @@
 /*********************
  *      INCLUDES
  *********************/
+#include <stdlib.h>
+
 #include "lv_obj.h"
 #include "lv_indev.h"
 #include "../misc/lv_anim.h"
@@ -28,9 +30,9 @@
 /**********************
  *  STATIC PROTOTYPES
  **********************/
-static void lv_obj_del_async_cb(void * obj);
+static void lv_obj_del_async_cb(void* obj);
 static void obj_del_core(lv_obj_t * obj);
-static lv_obj_tree_walk_res_t walk_core(lv_obj_t * obj, lv_obj_tree_walk_cb_t cb, void * user_data);
+static lv_obj_tree_walk_res_t walk_core(lv_obj_t * obj, lv_obj_tree_walk_cb_t cb, void* user_data);
 
 /**********************
  *  STATIC VARIABLES
@@ -73,7 +75,7 @@ void lv_obj_del(lv_obj_t * obj)
     }
 
     /*Handle if the active screen was deleted*/
-    if(act_scr_del)  {
+    if(act_scr_del) {
         LV_LOG_WARN("the active screen was deleted")
         disp->act_scr = NULL;
     }
@@ -159,8 +161,8 @@ void lv_obj_set_parent(lv_obj_t * obj, lv_obj_t * parent)
 
     /*Remove the object from the old parent's child list*/
     int32_t i;
-    for(i = lv_obj_get_child_id(obj); i <= (int32_t)lv_obj_get_child_cnt(old_parent) - 2; i++) {
-        old_parent->spec_attr->children[i] = old_parent->spec_attr->children[i+1];
+    for(i = lv_obj_get_index(obj); i <= (int32_t)lv_obj_get_child_cnt(old_parent) - 2; i++) {
+        old_parent->spec_attr->children[i] = old_parent->spec_attr->children[i + 1];
     }
     old_parent->spec_attr->child_cnt--;
     if(old_parent->spec_attr->child_cnt) {
@@ -202,7 +204,7 @@ void lv_obj_move_foreground(lv_obj_t * obj)
     lv_obj_t * parent = lv_obj_get_parent(obj);
 
     uint32_t i;
-    for(i = lv_obj_get_child_id(obj); i < lv_obj_get_child_cnt(parent) - 1; i++) {
+    for(i = lv_obj_get_index(obj); i < lv_obj_get_child_cnt(parent) - 1; i++) {
         parent->spec_attr->children[i] = parent->spec_attr->children[i + 1];
     }
     parent->spec_attr->children[lv_obj_get_child_cnt(parent) - 1] = obj;
@@ -220,7 +222,7 @@ void lv_obj_move_background(lv_obj_t * obj)
     lv_obj_t * parent = lv_obj_get_parent(obj);
 
     int32_t i;
-    for(i = lv_obj_get_child_id(obj); i > 0; i--) {
+    for(i = lv_obj_get_index(obj); i > 0; i--) {
         parent->spec_attr->children[i] = parent->spec_attr->children[i-1];
     }
     parent->spec_attr->children[0] = obj;
@@ -231,16 +233,52 @@ void lv_obj_move_background(lv_obj_t * obj)
     lv_obj_invalidate(parent);
 }
 
+void lv_obj_move_to_index(lv_obj_t * obj, int32_t index)
+{
+    LV_ASSERT_OBJ(obj, MY_CLASS);
+
+    const int32_t old_index = lv_obj_get_index(obj);
+
+    lv_obj_t * parent = lv_obj_get_parent(obj);
+
+    index = max(0, min(index, (int32_t ) lv_obj_get_child_cnt(parent) - 1));
+
+    if (index == old_index) return;
+
+    int32_t i = old_index;
+    if (index < old_index)
+    {
+        while (i > index)
+        {
+            parent->spec_attr->children[i] = parent->spec_attr->children[i - 1];
+            lv_event_send(parent, LV_EVENT_CHILD_CHANGED, parent->spec_attr->children[i]);
+            i--;
+        }
+    }
+    else
+    {
+        while (i < index)
+        {
+            parent->spec_attr->children[i] = parent->spec_attr->children[i + 1];
+            lv_event_send(parent, LV_EVENT_CHILD_CHANGED, parent->spec_attr->children[i]);
+            i++;
+        }
+    }
+    parent->spec_attr->children[index] = obj;
+    lv_event_send(parent, LV_EVENT_CHILD_CHANGED, obj);
+    lv_obj_invalidate(parent);
+}
+
 void lv_obj_swap(lv_obj_t * obj1, lv_obj_t * obj2)
 {
     LV_ASSERT_OBJ(obj1, MY_CLASS);
     LV_ASSERT_OBJ(obj2, MY_CLASS);
 
-    lv_obj_t* parent = lv_obj_get_parent(obj1);
-    lv_obj_t* parent2 = lv_obj_get_parent(obj2);
+    lv_obj_t * parent = lv_obj_get_parent(obj1);
+    lv_obj_t * parent2 = lv_obj_get_parent(obj2);
 
-    uint_fast32_t index1 = lv_obj_get_child_id(obj1);
-    uint_fast32_t index2 = lv_obj_get_child_id(obj2);
+    uint_fast32_t index1 = lv_obj_get_index(obj1);
+    uint_fast32_t index2 = lv_obj_get_index(obj2);
 
     parent->spec_attr->children[index1] = obj2;
     parent2->spec_attr->children[index2] = obj1;
@@ -249,10 +287,11 @@ void lv_obj_swap(lv_obj_t * obj1, lv_obj_t * obj2)
     lv_event_send(parent2, LV_EVENT_CHILD_CHANGED, obj1);
 
     lv_obj_invalidate(parent);
-    if( parent != parent2) {
+
+    if(parent != parent2)
+    {
         lv_obj_invalidate(parent2);
     }
-
     lv_group_swap_obj(obj1, obj2);
 }
 
@@ -260,12 +299,12 @@ void lv_obj_move_up(lv_obj_t * obj)
 {
     LV_ASSERT_OBJ(obj, MY_CLASS);
 
-    lv_obj_t* parent = lv_obj_get_parent(obj);
+    lv_obj_t * parent = lv_obj_get_parent(obj);
 
-    uint_fast32_t index = lv_obj_get_child_id(obj);
+    uint_fast32_t index = lv_obj_get_index(obj);
     if (index > 0)
     {
-        lv_obj_t* obj2 = parent->spec_attr->children[index - 1];
+        lv_obj_t * obj2 = parent->spec_attr->children[index - 1];
         parent->spec_attr->children[index - 1] = obj;
         parent->spec_attr->children[index] = obj2;
 
@@ -278,16 +317,17 @@ void lv_obj_move_up(lv_obj_t * obj)
     }
 }
 
-void lv_obj_move_down(lv_obj_t* obj)
+
+void lv_obj_move_down(lv_obj_t * obj)
 {
     LV_ASSERT_OBJ(obj, MY_CLASS);
 
-    lv_obj_t* parent = lv_obj_get_parent(obj);
+    lv_obj_t * parent = lv_obj_get_parent(obj);
 
-    uint_fast32_t index = lv_obj_get_child_id(obj);
+    uint_fast32_t index = lv_obj_get_index(obj);
     if (index < lv_obj_get_child_cnt(parent) - 1)
     {
-        lv_obj_t* obj2 = parent->spec_attr->children[index + 1];
+        lv_obj_t * obj2 = parent->spec_attr->children[index + 1];
         parent->spec_attr->children[index + 1] = obj;
         parent->spec_attr->children[index] = obj2;
 
@@ -309,10 +349,10 @@ lv_obj_t * lv_obj_get_screen(const lv_obj_t * obj)
 
     do {
         act_par = par;
-        par   = lv_obj_get_parent(act_par);
+        par = lv_obj_get_parent(act_par);
     } while(par != NULL);
 
-    return (lv_obj_t*)act_par;
+    return (lv_obj_t *)act_par;
 }
 
 lv_disp_t * lv_obj_get_disp(const lv_obj_t * obj)
@@ -370,7 +410,7 @@ uint32_t lv_obj_get_child_cnt(const lv_obj_t * obj)
     return obj->spec_attr->child_cnt;
 }
 
-uint32_t lv_obj_get_child_id(const lv_obj_t * obj)
+uint32_t lv_obj_get_index(const lv_obj_t * obj)
 {
     LV_ASSERT_OBJ(obj, MY_CLASS);
 
@@ -386,7 +426,7 @@ uint32_t lv_obj_get_child_id(const lv_obj_t * obj)
 }
 
 
-void lv_obj_tree_walk(lv_obj_t * start_obj, lv_obj_tree_walk_cb_t cb, void * user_data)
+void lv_obj_tree_walk(lv_obj_t * start_obj, lv_obj_tree_walk_cb_t cb, void* user_data)
 {
     walk_core(start_obj, cb, user_data);
 }
@@ -395,7 +435,7 @@ void lv_obj_tree_walk(lv_obj_t * start_obj, lv_obj_tree_walk_cb_t cb, void * use
  *   STATIC FUNCTIONS
  **********************/
 
-static void lv_obj_del_async_cb(void * obj)
+static void lv_obj_del_async_cb(void* obj)
 {
     LV_ASSERT_OBJ(obj, MY_CLASS);
 
@@ -415,10 +455,10 @@ static void obj_del_core(lv_obj_t * obj)
         child = lv_obj_get_child(obj, 0);
     }
 
-    lv_group_t * group = lv_obj_get_group(obj);
+    lv_group_t* group = lv_obj_get_group(obj);
 
     /*Reset all input devices if the object to delete is used*/
-    lv_indev_t * indev = lv_indev_get_next(NULL);
+    lv_indev_t* indev = lv_indev_get_next(NULL);
     while(indev) {
         if(indev->proc.types.pointer.act_obj == obj || indev->proc.types.pointer.last_obj == obj) {
             lv_indev_reset(indev, obj);
@@ -454,7 +494,7 @@ static void obj_del_core(lv_obj_t * obj)
     }
     /*Remove the object from the child list of its parent*/
     else {
-        uint32_t id = lv_obj_get_child_id(obj);
+        uint32_t id = lv_obj_get_index(obj);
         uint32_t i;
         for(i = id; i < obj->parent->spec_attr->child_cnt - 1; i++) {
             obj->parent->spec_attr->children[i] = obj->parent->spec_attr->children[i + 1];
@@ -468,7 +508,7 @@ static void obj_del_core(lv_obj_t * obj)
 }
 
 
-static lv_obj_tree_walk_res_t walk_core(lv_obj_t * obj, lv_obj_tree_walk_cb_t cb, void * user_data)
+static lv_obj_tree_walk_res_t walk_core(lv_obj_t * obj, lv_obj_tree_walk_cb_t cb, void* user_data)
 {
     lv_obj_tree_walk_res_t res = LV_OBJ_TREE_WALK_NEXT;